Create Meal time.py
Suppose that you’re in a country where it’s customary to eat breakfast between 7:00 and 8:00,
lunch between 12:00 and 13:00, and dinner between 18:00 and 19:00.
This program tells the user what to eat base on their time.

This is a program that prompts a user for a time and outputs whether it’s breakfast time,
lunchtime, or dinner time. If it’s not time for a meal,
depending on the real world time at their current location.